Refactoring to allow when condition plugins to not need p-m-d
Moving around various methods (and tweaking one of them to not need
pipeline-stage-step) so that plugins like
declarative-pipeline-when-conditions can avoid depending on pipeline-model-definition.
