+++ pipeline-model-definition/src/main/java/org/jenkinsci/plugins/pipeline/modeldefinition/when/impl/AllOfConditional.java
@@ -27,8 +27,8 @@ package org.jenkinsci.plugins.pipeline.modeldefinition.when.impl;
 import hudson.Extension;
 import org.codehaus.groovy.ast.expr.Expression;
 import org.jenkinsci.Symbol;
+import org.jenkinsci.plugins.pipeline.modeldefinition.DeclarativeExtensionUtils;
 import org.jenkinsci.plugins.pipeline.modeldefinition.ast.ModelASTWhenContent;
-import org.jenkinsci.plugins.pipeline.modeldefinition.parser.ASTParserUtils;
 import org.jenkinsci.plugins.pipeline.modeldefinition.when.DeclarativeStageConditional;
 import org.jenkinsci.plugins.pipeline.modeldefinition.when.DeclarativeStageConditionalDescriptor;
 import org.kohsuke.stapler.DataBoundConstructor;
@@ -55,7 +55,7 @@ public class AllOfConditional extends AbstractConditionalWithChildren<AllOfCondi
 
         @Override
         public Expression transformToRuntimeAST(@CheckForNull ModelASTWhenContent original) {
-            return ASTParserUtils.transformWhenContentToRuntimeAST(original);
+            return DeclarativeExtensionUtils.transformWhenContentToRuntimeAST(original);
         }
     }
 }
