+++ pipeline-model-definition/src/main/java/org/jenkinsci/plugins/pipeline/modeldefinition/when/impl/BranchConditional.java
@@ -29,8 +29,8 @@ import hudson.Extension;
 import org.apache.tools.ant.types.selectors.SelectorUtils;
 import org.codehaus.groovy.ast.expr.Expression;
 import org.jenkinsci.Symbol;
+import org.jenkinsci.plugins.pipeline.modeldefinition.DeclarativeExtensionUtils;
 import org.jenkinsci.plugins.pipeline.modeldefinition.ast.ModelASTWhenContent;
-import org.jenkinsci.plugins.pipeline.modeldefinition.parser.ASTParserUtils;
 import org.jenkinsci.plugins.pipeline.modeldefinition.when.DeclarativeStageConditional;
 import org.jenkinsci.plugins.pipeline.modeldefinition.when.DeclarativeStageConditionalDescriptor;
 import org.kohsuke.stapler.DataBoundConstructor;
@@ -75,7 +75,7 @@ public class BranchConditional extends DeclarativeStageConditional<BranchConditi
     public static class DescriptorImpl extends DeclarativeStageConditionalDescriptor<BranchConditional> {
         @Override
         public Expression transformToRuntimeAST(@CheckForNull ModelASTWhenContent original) {
-            return ASTParserUtils.transformWhenContentToRuntimeAST(original);
+            return DeclarativeExtensionUtils.transformWhenContentToRuntimeAST(original);
         }
     }
 }
